Test case: It's not possible to test that path_params is scrubed
Rails+kaminari currently has a small bug (PR incoming) where path_params
can be injected into url_for by an end user when they put it on into the
query params of a url.

However, when I tried to write a tests for rubygems.org to prevent
crashes, the tests themselves crash when path_params is passed as you
can see in this test case. I will link to this from the fix PR.